Apply audit findings: raw PTY channel, paused background terminals, palette, badge, fixes
