+++ src-tauri/src/lib.rs
@@ -17,6 +17,7 @@ use std::{
     time::{Duration, SystemTime},
 };
 use sysinfo::{ProcessRefreshKind, ProcessesToUpdate, System, UpdateKind};
+use tauri::ipc::{Channel, InvokeResponseBody};
 use tauri::{AppHandle, Emitter, Manager, State};
 use tauri_plugin_dialog::DialogExt;
 use tauri_plugin_updater::UpdaterExt;
@@ -181,6 +182,22 @@ fn send_notification(session_name: String, session_id: String) {
 #[tauri::command]
 fn send_notification(_session_name: String, _session_id: String) {}
 
+// The number of sessions waiting on the user, on the Dock or taskbar icon, where someone who has
+// switched to another app will see it. Windows has no badge count; it keeps the notification alone.
+#[tauri::command]
+fn set_attention_badge(app: AppHandle, count: u32) -> Result<(), String> {
+    let window = app.get_webview_window("main").ok_or("No main window")?;
+    #[cfg(windows)]
+    {
+        let _ = (window, count);
+        Ok(())
+    }
+    #[cfg(not(windows))]
+    window
+        .set_badge_count((count > 0).then_some(i64::from(count)))
+        .map_err(|error| error.to_string())
+}
+
 #[derive(Clone, Copy)]
 struct CodexProvider {
     id: &'static str,
@@ -213,7 +230,9 @@ const CODEX_PROVIDERS: [CodexProvider; 2] = [
 struct PtySession {
     child: Box<dyn Child + Send + Sync>,
     master: Box<dyn MasterPty + Send>,
-    writer: Box<dyn Write + Send>,
+    // Shared so a write takes only this session's lock: a paste into a busy child blocks on the tty,
+    // and holding the whole session table for that would stall every other tab's resize and stop.
+    writer: Arc<Mutex<Box<dyn Write + Send>>>,
     run_id: String,
     alive: Arc<AtomicBool>,
     agent_watch: Arc<AtomicU64>,
@@ -443,14 +462,6 @@ struct CodexServer(Mutex<CodexServerState>);
 struct CodexServerState {
     child: Option<std::process::Child>,
     endpoint: String,
-}
-
-#[derive(Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct PtyOutput {
-    session_id: String,
-    run_id: String,
-    data: Vec<u8>,
 }
 
 #[derive(Clone, Serialize)]
@@ -958,10 +969,10 @@ fn update_provider_session(
         if !is_provider_session_id(&provider_session_id) {
             return Err("Invalid provider session ID".into());
         }
-        if let Some(existing) = sessions.get(session_id) {
-            if existing == &provider_session_id {
-                return Ok(true);
-            }
+        if let Some(existing) = sessions.get(session_id)
+            && existing == &provider_session_id
+        {
+            return Ok(true);
         }
         // Claiming under the lock lets concurrent discoveries run without one stealing the other's session.
         if sessions
@@ -1211,18 +1222,17 @@ pub fn capture_claude_status(path: &str, activity_path: &str) -> Result<(), Stri
         ("five_hour", "Current session"),
         ("seven_day", "Current week"),
     ] {
-        if let Some(window) = input.get("rate_limits").and_then(|limits| limits.get(key)) {
-            if let Some(used_percent) = window
+        if let Some(window) = input.get("rate_limits").and_then(|limits| limits.get(key))
+            && let Some(used_percent) = window
                 .get("used_percentage")
                 .and_then(serde_json::Value::as_f64)
-            {
-                windows.push(UsageWindow {
-                    label: label.into(),
-                    used_percent,
-                    resets_at: window.get("resets_at").and_then(serde_json::Value::as_u64),
-                    window_minutes: None,
-                });
-            }
+        {
+            windows.push(UsageWindow {
+                label: label.into(),
+                used_percent,
+                resets_at: window.get("resets_at").and_then(serde_json::Value::as_u64),
+                window_minutes: None,
+            });
         }
     }
     let snapshot = UsageSnapshot {
@@ -1847,37 +1857,35 @@ fn codex_usage(
                 .and_then(serde_json::Value::as_str)
                 .unwrap_or("Codex");
             for (key, suffix) in [("primary", ""), ("secondary", " secondary")] {
-                if let Some(window) = bucket.get(key).filter(|value| !value.is_null()) {
-                    if let Some(used_percent) = window
+                if let Some(window) = bucket.get(key).filter(|value| !value.is_null())
+                    && let Some(used_percent) = window
                         .get("usedPercent")
                         .and_then(serde_json::Value::as_f64)
-                    {
-                        windows.push(UsageWindow {
-                            label: format!("{name}{suffix}"),
-                            used_percent,
-                            resets_at: window.get("resetsAt").and_then(serde_json::Value::as_u64),
-                            window_minutes: window
-                                .get("windowDurationMins")
-                                .and_then(serde_json::Value::as_u64),
-                        });
-                    }
+                {
+                    windows.push(UsageWindow {
+                        label: format!("{name}{suffix}"),
+                        used_percent,
+                        resets_at: window.get("resetsAt").and_then(serde_json::Value::as_u64),
+                        window_minutes: window
+                            .get("windowDurationMins")
+                            .and_then(serde_json::Value::as_u64),
+                    });
                 }
             }
         }
-    } else if let Some(window) = rates.and_then(|value| value.pointer("/rateLimits/primary")) {
-        if let Some(used_percent) = window
+    } else if let Some(window) = rates.and_then(|value| value.pointer("/rateLimits/primary"))
+        && let Some(used_percent) = window
             .get("usedPercent")
             .and_then(serde_json::Value::as_f64)
-        {
-            windows.push(UsageWindow {
-                label: "Codex".into(),
-                used_percent,
-                resets_at: window.get("resetsAt").and_then(serde_json::Value::as_u64),
-                window_minutes: window
-                    .get("windowDurationMins")
-                    .and_then(serde_json::Value::as_u64),
-            });
-        }
+    {
+        windows.push(UsageWindow {
+            label: "Codex".into(),
+            used_percent,
+            resets_at: window.get("resetsAt").and_then(serde_json::Value::as_u64),
+            window_minutes: window
+                .get("windowDurationMins")
+                .and_then(serde_json::Value::as_u64),
+        });
     }
     let context = responses
         .get(&3)
@@ -1940,16 +1948,15 @@ fn stop_codex_server(server: &CodexServer) {
     };
     #[cfg(unix)]
     {
-        for attempts in [500, 40] {
-            let _ = Command::new("kill")
-                .args(["-TERM", &child.id().to_string()])
-                .status();
-            for _ in 0..attempts {
-                if child.try_wait().ok().flatten().is_some() {
-                    return;
-                }
-                thread::sleep(Duration::from_millis(50));
+        // Runs while Lite quits, so the grace is short: two seconds for a clean exit, then a kill.
+        let _ = Command::new("kill")
+            .args(["-TERM", &child.id().to_string()])
+            .status();
+        for _ in 0..40 {
+            if child.try_wait().ok().flatten().is_some() {
+                return;
             }
+            thread::sleep(Duration::from_millis(50));
         }
         let _ = child.kill();
         let _ = child.wait();
@@ -2208,19 +2215,24 @@ fn load_api_keys(app: &AppHandle) -> HashMap<String, String> {
         .unwrap_or_default()
 }
 
+// The file is created owner-only rather than made so afterwards, so no save leaves it readable for a
+// moment between the rename and the change of mode.
 fn write_api_keys(app: &AppHandle, keys: &HashMap<String, String>) -> Result<(), String> {
     let path = api_keys_path(app)?;
-    write_atomic(
-        &path,
-        &serde_json::to_vec(keys).map_err(|error| error.to_string())?,
-    )?;
+    if let Some(directory) = path.parent() {
+        fs::create_dir_all(directory).map_err(|error| error.to_string())?;
+    }
+    let contents = serde_json::to_vec(keys).map_err(|error| error.to_string())?;
+    let mut options = fs::OpenOptions::new();
+    options.write(true).create(true).truncate(true);
     #[cfg(unix)]
     {
-        use std::os::unix::fs::PermissionsExt;
-        fs::set_permissions(&path, fs::Permissions::from_mode(0o600))
-            .map_err(|error| error.to_string())?;
+        use std::os::unix::fs::OpenOptionsExt;
+        options.mode(0o600);
     }
-    Ok(())
+    AtomicFile::new(&path, AllowOverwrite)
+        .write_with_options(|file| file.write_all(&contents), options)
+        .map_err(|error| error.to_string())
 }
 
 // macOS keeps Claude Code's credentials in the login keychain; other platforms write a file. Presence is
@@ -3191,6 +3203,7 @@ async fn spawn_session(
     roots: State<'_, Roots>,
     provider_sessions: State<'_, ProviderSessions>,
     codex_server: State<'_, CodexServer>,
+    output: Channel<InvokeResponseBody>,
     session_id: String,
     run_id: String,
     root_id: String,
@@ -3247,16 +3260,16 @@ async fn spawn_session(
             uuid::Uuid::new_v4().to_string()
         });
     }
-    if !signing_in && (agent == "codex" || agent == "kimi") {
-        if let Some(saved_provider_session_id) = provider_sessions
+    if !signing_in
+        && (agent == "codex" || agent == "kimi")
+        && let Some(saved_provider_session_id) = provider_sessions
             .0
             .lock()
             .map_err(|error| error.to_string())?
             .get(&session_id)
             .cloned()
-        {
-            provider_session_id = Some(saved_provider_session_id);
-        }
+    {
+        provider_session_id = Some(saved_provider_session_id);
     }
     // An id the tab already holds is recorded here too, so no other tab claims that same session.
     if let Some(known) = provider_session_id.as_deref() {
@@ -3295,13 +3308,12 @@ async fn spawn_session(
             pixel_height: 0,
         })
         .map_err(|error| error.to_string())?;
+    // An app server that cannot answer costs the check, not the session: the tab opens on the recorded
+    // thread and Codex itself says whether it can resume, as the discovery below already assumes.
     if !signing_in
         && agent == "codex"
-        && provider_session_id.is_some()
-        && !codex_thread_resumable(
-            &codex_server,
-            provider_session_id.as_deref().expect("restored above"),
-        )?
+        && let Some(thread_id) = provider_session_id.as_deref()
+        && !codex_thread_resumable(&codex_server, thread_id).unwrap_or(true)
     {
         update_provider_session(&app, &provider_sessions, &session_id, None)?;
         provider_session_id = None;
@@ -3392,7 +3404,7 @@ async fn spawn_session(
         PtySession {
             child,
             master: pair.master,
-            writer,
+            writer: Arc::new(Mutex::new(writer)),
             run_id: run_id.clone(),
             alive: Arc::new(AtomicBool::new(true)),
             agent_watch: Arc::new(AtomicU64::new(0)),
@@ -3404,20 +3416,15 @@ async fn spawn_session(
     let event_run_id = run_id.clone();
     let output_app = app.clone();
     thread::spawn(move || {
+        // Bytes go to the page as bytes over the launch's own channel; an event would spell each one
+        // out as a JSON number.
         let mut buffer = [0_u8; 8192];
         while let Ok(count) = reader.read(&mut buffer) {
             if count == 0 {
                 break;
             }
-            if output_app
-                .emit(
-                    "pty-output",
-                    PtyOutput {
-                        session_id: event_session_id.clone(),
-                        run_id: event_run_id.clone(),
-                        data: buffer[..count].to_vec(),
-                    },
-                )
+            if output
+                .send(InvokeResponseBody::Raw(buffer[..count].to_vec()))
                 .is_err()
             {
                 break;
@@ -3455,6 +3462,9 @@ async fn spawn_session(
         let discovery_session_id = session_id.clone();
         let discovery_agent = agent.clone();
         thread::spawn(move || {
+            // The id appears with the first turn, which may be minutes away, so an idle tab is asked
+            // less and less often rather than every second for as long as it stays open.
+            let mut wait = Duration::from_secs(1);
             loop {
                 let candidates = if discovery_agent == "kimi" {
                     Ok(kimi_current_session(&discovery_app, &cwd)
@@ -3486,7 +3496,8 @@ async fn spawn_session(
                 if !running {
                     break;
                 }
-                thread::sleep(Duration::from_secs(1));
+                thread::sleep(wait);
+                wait = (wait * 2).min(Duration::from_secs(10));
             }
         });
     }
@@ -3626,20 +3637,22 @@ fn watch_shell_agent(
 }
 
 #[tauri::command]
-fn write_session(
-    sessions: State<Sessions>,
+async fn write_session(
+    sessions: State<'_, Sessions>,
     session_id: String,
     data: Vec<u8>,
 ) -> Result<(), String> {
-    let mut sessions = sessions.0.lock().map_err(|error| error.to_string())?;
-    let session = sessions
-        .get_mut(&session_id)
-        .ok_or("Session is not running")?;
-    session
-        .writer
-        .write_all(&data)
-        .map_err(|error| error.to_string())?;
-    session.writer.flush().map_err(|error| error.to_string())
+    let writer = {
+        let sessions = sessions.0.lock().map_err(|error| error.to_string())?;
+        sessions
+            .get(&session_id)
+            .ok_or("Session is not running")?
+            .writer
+            .clone()
+    };
+    let mut writer = writer.lock().map_err(|error| error.to_string())?;
+    writer.write_all(&data).map_err(|error| error.to_string())?;
+    writer.flush().map_err(|error| error.to_string())
 }
 
 #[tauri::command]
@@ -5024,7 +5037,7 @@ fn stop_runtime(app: &AppHandle) {
 // A local build names the commit it came from instead of the release version it would otherwise be
 // mistaken for. It follows main directly so fixes can be used before release assets exist.
 #[tauri::command]
-fn local_update() -> Result<Option<String>, String> {
+async fn local_update() -> Result<Option<String>, String> {
     let built = option_env!("LITE_COMMIT").ok_or("This is not a local build")?;
     let repo = option_env!("LITE_REPO").ok_or("This build did not record where it came from")?;
     let git = resolve_executable("git").unwrap_or_else(|| "git".into());
@@ -5198,6 +5211,7 @@ pub fn run() {
             Ok(())
         })
         .invoke_handler(tauri::generate_handler![
+            set_attention_badge,
             choose_directory,
             follow_directory,
             github_items,
