+++ src/components/ui/scroll-area.tsx
@@ -24,10 +24,11 @@ function ScrollBar({ className, orientation = "vertical", ...props }: ScrollArea
   return (
     <ScrollAreaPrimitive.Scrollbar
       data-slot="scroll-area-scrollbar"
-      data-orientation={orientation}
       orientation={orientation}
       className={cn(
-        "flex touch-none p-px transition-colors select-none data-horizontal:h-2.5 data-horizontal:flex-col data-horizontal:border-t data-horizontal:border-t-transparent data-vertical:h-full data-vertical:w-2.5 data-vertical:border-l data-vertical:border-l-transparent",
+        // Base UI reports the axis as data-orientation. The bars live in narrow panels, so they are slimmer
+        // than the component's default.
+        "flex touch-none p-px transition-colors select-none data-[orientation=horizontal]:h-1.5 data-[orientation=horizontal]:flex-col data-[orientation=horizontal]:border-t data-[orientation=horizontal]:border-t-transparent data-[orientation=vertical]:h-full data-[orientation=vertical]:w-1.5 data-[orientation=vertical]:border-l data-[orientation=vertical]:border-l-transparent",
         className,
       )}
       {...props}
